Merge 3/2

+++ unit/unit.go
@@ -15,13 +15,17 @@ import (
 
 	"github.com/gruntwork-io/terratest/modules/random"
 	"github.com/gruntwork-io/terratest/modules/terraform"
+	"github.com/hashicorp/terraform-json"
 )
 
 // ResourceDescription Identifies mappings between resources and attributes
 type ResourceDescription map[string]map[string]interface{}
 
 // TerraformPlanValidation A function that can run an assertion over a terraform plan
-type TerraformPlanValidation func(goTest *testing.T, plan TerraformPlan)
+type TerraformPlanValidation func(goTest *testing.T, plan tfjson.Plan)
+
+// TerraformCommandStdoutValidation A function that can run an assertion over a terraform command output and exit code
+type TerraformCommandStdoutValidation func(goTest *testing.T, output string, err error)
 
 // UnitTestFixture Holds metadata required to execute a unit test against a test against a terraform template
 type UnitTestFixture struct {
@@ -31,7 +35,8 @@ type UnitTestFixture struct {
 	ExpectedResourceCount int // Expected # of resources that Terraform should create
 	// map of maps specifying resource <--> attribute <--> attribute value mappings
 	ExpectedResourceAttributeValues ResourceDescription
-	PlanAssertions                  []TerraformPlanValidation // user-defined plan assertions
+	PlanAssertions                  []TerraformPlanValidation          // user-defined plan assertions
+	CommandStdoutAssertions         []TerraformCommandStdoutValidation // user-defined command output assertions
 }
 
 // RunUnitTests Executes terraform lifecycle events and verifies the correctness of the resulting terraform.
@@ -54,17 +59,36 @@ func RunUnitTests(fixture *UnitTestFixture) {
 		terraform.FormatArgs(fixture.TfOptions, "workspace", "show")...)
 
 	terraform.WorkspaceSelectOrNew(fixture.GoTest, fixture.TfOptions, workspaceName)
-	defer terraform.RunTerraformCommand(fixture.GoTest, fixture.TfOptions, "workspace", "delete", workspaceName)
+	if startingWorkspaceName != workspaceName {
+		defer terraform.RunTerraformCommand(fixture.GoTest, fixture.TfOptions, "workspace", "delete", workspaceName)
+	}
 	defer terraform.WorkspaceSelectOrNew(fixture.GoTest, fixture.TfOptions, startingWorkspaceName)
 
 	tfPlanFilePath := filepath.FromSlash(fmt.Sprintf("%s/%s.plan", os.TempDir(), random.UniqueId()))
-	terraform.RunTerraformCommand(
+	defer os.Remove(tfPlanFilePath)
+
+	output, err := terraform.RunTerraformCommandE(
 		fixture.GoTest,
 		fixture.TfOptions,
 		terraform.FormatArgs(fixture.TfOptions, "plan", "-input=false", "-out", tfPlanFilePath)...)
-	defer os.Remove(tfPlanFilePath)
+	if err != nil && fixture.CommandStdoutAssertions == nil {
+		fixture.GoTest.Fatal(err)
+	}
+	if fixture.CommandStdoutAssertions != nil {
+		validateTerraformCommandStdout(fixture, output, err)
+	}
+	if err == nil {
+		validateTerraformPlanFile(fixture, tfPlanFilePath)
+	}
+}
 
-	validateTerraformPlanFile(fixture, tfPlanFilePath)
+// Validate a failed terraform command output and error
+func validateTerraformCommandStdout(fixture *UnitTestFixture, output string, err error) {
+	if fixture.CommandStdoutAssertions != nil {
+		for _, assertion := range fixture.CommandStdoutAssertions {
+			assertion(fixture.GoTest, output, err)
+		}
+	}
 }
 
 // Validates a terraform plan file based on the test fixture. The following validations are made:
@@ -102,7 +126,7 @@ func validateTerraformPlanFile(fixture *UnitTestFixture, tfPlanFilePath string) 
 	}
 }
 
-func parseTerraformPlan(fixture *UnitTestFixture, filePath string) TerraformPlan {
+func parseTerraformPlan(fixture *UnitTestFixture, filePath string) tfjson.Plan {
 	// Note: when the PR linked below is merged and the new build is used by this codebase,
 	// we can leverage Terratest to run this for us. Currently with large plan outputs,
 	// a buffer overflow will happen in Terratest because the default max character limit
@@ -123,7 +147,7 @@ func parseTerraformPlan(fixture *UnitTestFixture, filePath string) TerraformPlan
 	}
 
 	fmt.Println("Got terraform plan...", string(jsonBytes))
-	var plan TerraformPlan
+	var plan tfjson.Plan
 	jsonErr := json.Unmarshal(jsonBytes, &plan)
 	if jsonErr != nil {
 		fixture.GoTest.Fatal(jsonErr)
@@ -132,14 +156,14 @@ func parseTerraformPlan(fixture *UnitTestFixture, filePath string) TerraformPlan
 }
 
 // Validates that the plan is not empty
-func validatePlanNotEmpty(t *testing.T, plan TerraformPlan) {
+func validatePlanNotEmpty(t *testing.T, plan tfjson.Plan) {
 	if len(plan.ResourceChanges) == 0 {
 		t.Fatalf("Plan diff was unexpectedly empty")
 	}
 }
 
 // Validates that the plan has the correct number of resources in it
-func validatePlanResourceCount(t *testing.T, fixture *UnitTestFixture, plan TerraformPlan) {
+func validatePlanResourceCount(t *testing.T, fixture *UnitTestFixture, plan tfjson.Plan) {
 	if len(plan.ResourceChanges) != fixture.ExpectedResourceCount {
 		t.Fatalf(
 			"Plan unexpectedly had %d resources instead of %d", len(plan.ResourceChanges), fixture.ExpectedResourceCount)
@@ -147,9 +171,9 @@ func validatePlanResourceCount(t *testing.T, fixture *UnitTestFixture, plan Terr
 }
 
 // Validates that the plan is not executing any destructive actions
-func validatePlanHasNoDeletes(t *testing.T, plan TerraformPlan) {
+func validatePlanHasNoDeletes(t *testing.T, plan tfjson.Plan) {
 	// a unit test should never create a destructive action like deleting a resource
-	allowedActions := map[string]bool{"create": true, "read": true}
+	allowedActions := map[tfjson.Action]bool{tfjson.ActionCreate: true, tfjson.ActionRead: true}
 	for _, resource := range plan.ResourceChanges {
 		for _, action := range resource.Change.Actions {
 			if !allowedActions[action] {
@@ -161,7 +185,7 @@ func validatePlanHasNoDeletes(t *testing.T, plan TerraformPlan) {
 
 // verifies that the attribute value mappings for each resource specified by the client exist
 // as a subset of the actual values defined in the terraform plan.
-func validatePlanResourceKeyValues(t *testing.T, fixture *UnitTestFixture, plan TerraformPlan) {
+func validatePlanResourceKeyValues(t *testing.T, fixture *UnitTestFixture, plan tfjson.Plan) {
 	theRealPlanAsMap := planToMap(plan)
 	theExpectedPlanAsMap := resourceDescriptionToMap(fixture.ExpectedResourceAttributeValues)
 
@@ -171,7 +195,7 @@ func validatePlanResourceKeyValues(t *testing.T, fixture *UnitTestFixture, plan 
 }
 
 // transforms the output of `terraform show -json <planfile>` into a generic map
-func planToMap(plan TerraformPlan) map[string]interface{} {
+func planToMap(plan tfjson.Plan) map[string]interface{} {
 	mp := make(map[string]interface{})
 	for _, resource := range plan.ResourceChanges {
 		mp[resource.Address] = resource.Change.After
